Migrate legacy start page to Symfony controller and Twig template
