+++ app/Resources/views/start/news.html.twig
@@ -1,0 +1,7 @@
+<h3>Aktuelle Neuigkeiten <span class="more">(<a href="{{ news_archive_url }}">mehr</a>)</span></h3><a
+        href="{{ news_feed_url }}" class="rss-icon"><img src="{{ asset('style/rss.png') }}" alt="RSS Feed"/></a>
+{% for entry in news_feed %}
+    <h4><a href="{{ entry.link }}">{{ entry.title }}</a></h4>
+    <p class="date">{{ entry.updated | date('d.m.Y') }}</p>
+    {{ entry.summary | raw }}
+{% endfor %}
