+++ app/Resources/views/start/recent_packages.html.twig
@@ -1,0 +1,17 @@
+<h3>Aktuelle Pakete <span class="more">(<a href="{{ path('app_legacy_page', {'page': 'Packages'}) }}">mehr</a>)</span>
+</h3><a
+        href="{{ path('app_legacy_page', {'page': 'GetRecentPackages'}) }}" class="rss-icon"><img
+            src="{{ asset('style/rss.png') }}" alt="RSS Feed"/></a>
+<table>
+    {% for package in packages %}
+        <tr{% if package.testing == 1 %} class="testing"{% endif %}>
+            <td class="pkgname"><a href="{{ path('app_legacy_page', {
+                    'page': 'PackageDetails',
+                    'repo': package.repository,
+                    'arch': package.architecture,
+                    'pkgname': package.name
+                }) }}">{{ package.name }}</a></td>
+            <td class="pkgver">{{ package.version }}</td>
+        </tr>
+    {% endfor %}
+</table>
