--- src/archportal/l10n/de_DE/StartSidebar.html
@@ -1,39 +1,0 @@
-<h4>Dokumentation</h4>
-<ul>
-    <li><a href="https://wiki.archlinux.de/">Wiki</a></li>
-    <li><a href="https://wiki.archlinux.de/title/Arch_Install_Scripts">Arch Install Scripts</a></li>
-    <li><a href="https://wiki.archlinux.de/title/Anleitung_f%C3%BCr_Einsteiger">Anleitung für Einsteiger</a></li>
-</ul>
-<h4>Gemeinschaft</h4>
-<ul>
-    <li><a href="https://planet.archlinux.de/">Planet archlinux.de</a></li>
-    <li><a href="https://www.archlinux.org/">Archlinux.org</a></li>
-    <li><a href="https://wiki.archlinux.org/index.php/International_Communities">Internationale Gemeinschaft</a></li>
-</ul>
-<h4>Unterstützung</h4>
-<ul>
-    <li><a href="https://wiki.archlinux.de/title/Spenden">Spenden (archlinux.de)</a></li>
-    <li><a href="https://www.archlinux.org/donate/">Spenden (international)</a></li>
-</ul>
-<h4>Entwicklung</h4>
-<ul>
-    <li><a href="?page=Packages">Pakete</a></li>
-    <li><a href="https://www.archlinux.org/packages/differences/">Architektur-Unterschiede</a></li>
-    <li><a href="https://aur.archlinux.de/">AUR</a></li>
-    <li><a href="https://bugs.archlinux.org/">Bug Tracker</a></li>
-    <li><a href="https://www.archlinux.org/svn/">SVN Repositories</a></li>
-    <li><a href="https://projects.archlinux.org/">Projekte in Git</a></li>
-    <li><a href="https://projects.archlinux.de/">archlinux.de Projekte in Git</a></li>
-    <li><a href="https://wiki.archlinux.org/index.php/DeveloperWiki">Entwickler-Wiki</a></li>
-</ul>
-<h4>Informationen</h4>
-<ul>
-    <li><a href="https://wiki.archlinux.de/title/%C3%9Cber_Arch_Linux">über Arch Linux</a></li>
-    <li><a href="https://wiki.archlinux.de/title/Download">Arch herunterladen</a></li>
-    <li><a href="https://wiki.archlinux.de/title/Arch_in_den_Medien">Arch in den Medien</a></li>
-    <li><a href="https://www.archlinux.org/art/">Logos</a></li>
-    <li><a href="https://www.archlinux.org/developers/">Entwickler</a></li>
-    <li><a href="https://www.archlinux.org/trustedusers/">Trusted Users</a></li>
-    <li><a href="https://www.archlinux.org/fellows/">Ehemalige</a></li>
-    <li><a href="?page=MirrorStatus">Mirror-Status</a></li>
-</ul>
