--- src/archportal/l10n/de_DE/StartWelcome.html
@@ -1,14 +1,0 @@
-<h2>Willkommen bei Arch Linux</h2>
-<p>
-    <strong>Arch Linux</strong> ist eine <em>flexible</em> und <em>leichtgewichtige</em> Distribution für jeden
-    erdenklichen Einsatz-Zweck. Ein einfaches Grundsystem kann nach den Bedürfnissen des jeweiligen Nutzers nahezu
-    beliebig erweitert werden.
-</p>
-<p>
-    Nach einem gleitenden Release-System bieten wir zur Zeit vorkompilierte Pakete für die <code>i686</code>- und <code>x86_64</code>-Architekturen
-    an. Zusätzliche Werkzeuge ermöglichen zudem den schnellen Eigenbau von Paketen.
-</p>
-<p>
-    Arch Linux ist daher eine perfekte Distribution für erfahrene Anwender &mdash; und solche, die es werden wollen...
-</p>
-<p class="readmore"><a href="https://wiki.archlinux.de/title/%C3%9Cber_Arch_Linux">mehr über Arch Linux</a></p>
