--- src/archportal/l10n/en_US/StartSidebar.html
@@ -1,8 +1,0 @@
-<h4>Development</h4>
-<ul>
-    <li><a href="?page=Packages">Pakete</a></li>
-</ul>
-<h4>Tools</h4>
-<ul>
-    <li><a href="?page=MirrorStatus">Mirror-Status</a></li>
-</ul>
