--- src/archportal/l10n/en_US/StartWelcome.html
@@ -1,23 +1,0 @@
-<h2>A simple, lightweight distribution</h2>
-
-<p>You've reached the website for <strong>Arch Linux</strong>, a
-    lightweight and flexible Linux® distribution that tries to Keep It
-    Simple.</p>
-
-<p>Currently we have official packages optimized for the i686 and
-    x86-64 architectures. We complement our official package sets with a
-    <a href="https://aur.archlinux.org" title="Arch User Repository (AUR)">
-        community-operated package repository</a> that grows in size and
-    quality each and every day.</p>
-
-<p>Our strong community is diverse and helpful, and we pride ourselves
-    on the range of skillsets and uses for Arch that stem from it. Please
-    check out our <a href="https://bbs.archlinux.org" title="Arch Forums">forums</a>
-    and <a href="http://mailman.archlinux.org/mailman/listinfo/"
-           title="Arch Mailing Lists">mailing lists</a>
-    to get your feet wet. Also glance through our <a href="https://wiki.archlinux.org"
-                                                     title="Arch Wiki">wiki</a>
-    if you want to learn more about Arch.</p>
-
-<p class="readmore"><a href="/about/"
-                       title="Learn more about Arch Linux">Learn more...</a></p>
