+++ web/style/arch.css
@@ -219,7 +219,7 @@ h5 {
     padding: 0;
 }
 
-#archnavbar.anb-start ul li#anb-start a {
+#archnavbar.anb-app_start_index ul li#anb-start a {
     color: white !important;
 }
